+++ src/main/java/org/jabref/logic/pdf/search/indexing/PdfIndexer.java
@@ -59,9 +59,8 @@ public class PdfIndexer {
      */
     public void createIndex() {
         // Create new index by creating IndexWriter but not writing anything.
-        try {
-            IndexWriter indexWriter = new IndexWriter(directoryToIndex, new IndexWriterConfig(new EnglishStemAnalyzer()).setOpenMode(IndexWriterConfig.OpenMode.CREATE));
-            indexWriter.close();
+        try (IndexWriter indexWriter = new IndexWriter(directoryToIndex, new IndexWriterConfig(new EnglishStemAnalyzer()).setOpenMode(IndexWriterConfig.OpenMode.CREATE))) {
+            // empty comment for checkstyle
         } catch (IOException e) {
             LOGGER.warn("Could not create new Index!", e);
         }
@@ -119,8 +118,7 @@ public class PdfIndexer {
         try (IndexWriter indexWriter = new IndexWriter(
                 directoryToIndex,
                 new IndexWriterConfig(
-                        new EnglishStemAnalyzer()).setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND))
-        ) {
+                        new EnglishStemAnalyzer()).setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND))) {
             if (!entry.getFiles().isEmpty()) {
                 indexWriter.deleteDocuments(new Term(SearchFieldConstants.PATH, linkedFile.getLink()));
             }
@@ -189,8 +187,7 @@ public class PdfIndexer {
         }
         try {
             // Check if a document with this path is already in the index
-            try {
-                IndexReader reader = DirectoryReader.open(directoryToIndex);
+            try (IndexReader reader = DirectoryReader.open(directoryToIndex)) {
                 IndexSearcher searcher = new IndexSearcher(reader);
                 TermQuery query = new TermQuery(new Term(SearchFieldConstants.PATH, linkedFile.getLink()));
                 TopDocs topDocs = searcher.search(query, 1);
@@ -205,19 +202,18 @@ public class PdfIndexer {
                         return;
                     }
                 }
-                reader.close();
             } catch (IndexNotFoundException e) {
                 // if there is no index yet, don't need to check anything!
             }
             // If no document was found, add the new one
             Optional<List<Document>> pages = new DocumentReader(entry, filePreferences).readLinkedPdf(this.databaseContext, linkedFile);
             if (pages.isPresent()) {
-                IndexWriter indexWriter = new IndexWriter(directoryToIndex,
-                        new IndexWriterConfig(
-                                new EnglishStemAnalyzer()).setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND));
-                indexWriter.addDocuments(pages.get());
-                indexWriter.commit();
-                indexWriter.close();
+                try (IndexWriter indexWriter = new IndexWriter(directoryToIndex,
+                                                               new IndexWriterConfig(
+                                                                                     new EnglishStemAnalyzer()).setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND))) {
+                    indexWriter.addDocuments(pages.get());
+                    indexWriter.commit();
+                }
             }
         } catch (IOException e) {
             LOGGER.warn("Could not add the document {} to the index!", linkedFile.getLink(), e);